Allow selection of statistics to be shipped.
This patch extends bear:get_statistics/1 to allow the user to specify
which statistics are desired. This defaults to the full set available.

This patch also removes the scan_res2 tuple and makes the second-pass
calculation and list sorting lazy - if a user doesn't request variance,
skew, etc., then the second pass won't be done; similarly with
percentiles and sorting.

I'm not thrilled with the massive bear:compute_statistics/4 definition,
but the alternatives I considered mostly involved nasty function exports
and unnatural sigs.

Breaking changes:
  1. The median stat has been replaced by a percentile at 0.5
  2. Percentiles are now specified as floats